Tussentijdse commit, SetWidth en SetHeight aan interface toegevoegd, default value voor hdlr_soun en hdlr_vide toegevoegd, default value voor mvhd->NextTrackID() aangepast zodat deze geldig is

+++ MP4/box_tkhd.h
@@ -11,8 +11,8 @@ class Box_tkhd {
     void SetCreationTime( uint32_t TimeStamp = 0 );
     void SetModificationTime( uint32_t TimeStamp = 0 );
     void SetDurationTime( uint32_t TimeUnits = 0 );
-    void SetWidth( uint32_t Width = 0 );
-    void SetHeight( uint32_t Height = 0 );
+    void SetWidth( uint16_t Width = 0 );
+    void SetHeight( uint16_t Height = 0 );
     void SetFlags( bool Bit0 = true, bool Bit1 = true, bool Bit2 = true );
     void SetVersion( uint32_t Version = 0 );
     void SetTrackID( uint32_t TrackID = 0 );
